requirements.txt
# Core data science stack
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Data serialization
joblib>=1.2.0

# Configuration
PyYAML>=6.0

# Jupyter notebooks
notebook>=6.5.0
ipykernel>=6.20.0

# Testing
pytest>=7.2.0

# Optional: XGBoost (uncomment or install separately)
# xgboost>=1.7.0

# Optional: SHAP interpretability (uncomment or install separately)
# shap>=0.42.0
